Protos

+++ Shared/Models/Article.cs
@@ -6,7 +6,7 @@
         public virtual string ArticleNumber { get; set; }
         public virtual string Name { get; set; }
         public virtual string Description { get; set; }
-        public virtual decimal Price { get; set; }
+        public virtual double Price { get; set; }
         public virtual int Version { get; set; }
     }
 }
